requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
firebase-admin==6.5.0
chromadb==1.5.5
sentence-transformers==3.0.1
anthropic==0.30.0
google-generativeai==0.7.2
pdfplumber==0.11.1
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84
langdetect==1.0.9
requests==2.32.3
PyJWT==2.9.0
email-validator==2.2.0
fpdf2==2.8.2
